Added a helper function to disconnect the camera. It sets the camera connected flag to false and unreferences the camera handle. It is now called from the connect camera function when the camera is not found (admittedly this causes a redundant ction of setting the connected flag false but the idea is that this function will be called from other places also.
Added the empty context to be used by the connection mechanism but havint modified the connection function to use it yet as it is still useful for testing.

Also introduced the helper function to fetch a setting value from the camera. With the context now hadling error and status reporting this function is now greatly simplified